utils cleanup & refactors
