+++ package.json
@@ -29,6 +29,7 @@
     "@kubernetes/client-node": "^0.20.0",
     "decimal.js": "^10.4.3",
     "express": "^4.18.2",
+    "mime": "^1.6.0",
     "pg": "^8.11.3",
     "swagger-ui-express": "^5.0.0",
     "ts-node": "^10.9.1",
